+++ scripts/send.js
@@ -23,7 +23,7 @@ const formHtml = fs.readFileSync(path.join(__dirname, "./resources/simple-form.h
 
 // Create a transport with MailDev's default receiving port
 const transporter = nodemailer.createTransport({
-  port: 1025,
+  port: process.env.MAILDEV_SMTP_PORT || 1025,
 });
 
 // Messages list